Crate splay

Contains an implementation of splay trees where each node has a key/value pair to be used in maps and sets. The only requirement is that the key must implement the Ord trait.

Example

use splay::SplayMap;

let mut map = SplayMap::new();
map.insert("foo", "bar");
map.insert("hello", "world");
map.insert("splay", "tree");

for (k, v) in map.into_iter() {
    println!("{} => {}", k, v);
}

Reexports

pub use self::set::SplaySet;

pub use self::map::SplayMap;

Modules

map

set