Struct splay::set::SplaySet

pub struct SplaySet<T: Ord> { /* fields omitted */ }

Methods

impl<T: Ord> SplaySet<T>

fn new() -> SplaySet<T>

Creates a new empty set

fn into_iter(self) -> IntoIter<T>

Moves all values out of this set, transferring ownership to the given iterator.

fn len(&self) -> usize

fn is_empty(&self) -> bool

fn clear(&mut self)

fn contains(&self, t: &T) -> bool

Return true if the set contains a value

fn insert(&mut self, t: T) -> bool

Add a value to the set. Return true if the value was not already present in the set.

fn remove(&mut self, t: &T) -> bool

Remove a value from the set. Return true if the value was present in the set.